Use broadcast mechanism for killing servers.
I think this is cleaner. I imagine one of the reasons I did this
originally was because I thought if two people try to kill the server,
it will fail with a panic(), but only one thing can ever call close on
the server and that's the monitorServers() goroutine. So having said
that it should be able to keep enough state to know better. Rather than
having that be an operation that can occur multiple times (read on a
channel to kill the server) it seems like this is better.

- Clean up some tests that would fail randomly