Added command to shorten url. Added config to Google().
Update to lastest version of aarondl/query is required.
Google() now includes a shortened url to the results page if a Google
API key is present in the config of the query module. Google()
continues to work without an API key, but will not display any link to
the results page. The bot accepts the command "shorten" and returns a
shortened version of the supplied url if possible.

Query module config update: Added "GoogleApiKey".