Fix leaking passwords to extensions
In the past all commands and events were executed in parallel. The
problem with this is that sensitive bot commands (like auth) were also
handled by user and remote extension created event handlers for PRIVMSG,
meaning that the AUTH command could be seen by anyone with access to
create an extension.

In order to fix this, the core commands have been given their own
dispatcher & dispatcher core. These core commands (if present) are run
first and if any of them are activated by the dispatched message, the
user-created event handlers and command event handlers are never fired.